10/10/2019 Update
Created a custom database. Items are allowed to be created. User inputs title, content, date and time. Items are saved to a database. The application can access the database and update the list view and display the items on the screen. Currently implementing editing. Editing has access to the database and can set values, but currently working on Text view edits.